Patch tous les bug !!!!!
Bon alors déjà les if en "?" c'est VarTruc ? vrais : faux et pas faux vrais
Pour se qui est des Behaviour c'était des seconde et pas des milliseconde ;-;
Il avait quoi d'autre à oui j'ai oubliez de testée si des variable sont null ;(
Bon des erreur conne qu'il ne faux pas refaire mais qui son toujours les plus récurent. Voila elle belle exemple pour toi rootin aux raison de t'est possible bug. Je parle je parle mais j'ai encore des truc a faire moi. A plus :)

+++ VT-Reférance/Method/Methods.cs
@@ -150,13 +150,13 @@ namespace VT_Referance.Method
         /// <returns></returns>
         public static int GetVoltage()
         {
-            int totalvoltagefloat = 0;
+            float totalvoltagefloat = 0;
             foreach (var i in Generator079.Generators)
             {
-                totalvoltagefloat += (int)i.localVoltage;
+                totalvoltagefloat += i.localVoltage;
             }
             totalvoltagefloat *= 1000;
-            return totalvoltagefloat;
+            return (int)totalvoltagefloat;
         }
 
     }
